[A11y] Implement details-roles for popover targets of aria-details
This supports a11y for the popover API, which is still behind a flag.
See here for some examples:
https://developer.chrome.com/blog/pop-ups-theyre-making-a-resurgence

This object attribute provides hints describing what aria-details is pointing to. It is optional for the AT to consume this information to potentially provide useful information as the user navigates to the source of the relation.

Bug: none
Change-Id: I3766ae1907eccef20186067f33c8daaae06c29a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4240765
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1105282}