float: Fix full restore not restoring floated windows in tablet
Tablet mode has an overrided window state, which is usually
normal window state. We use this so windows that are forced maximized
in tablet mode aren't restored as maximized if in clamshell mode. Float
state should be exempt since it is a user chosen state.

Test: manual, ash_unittests WindowRestoreControllerTest.TabletFloatWindow
Change-Id: I0dc1f2597af0281b2220700feed80da1bfebb430
Fixed: b/268381072
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4249942
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Daniel Andersson <dandersson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105309}