[MediaFoundation] Fix lag when scrolling page during playback
OS: Win
When scrolling page during a playback, output rectangle of the video
will change, this change will call Media Foundations engine to update
the video rectangle. A scroll will generate many of these calls and
affect performance since statistics of frame presentation is reset when
a new video rectangle is called. Thus causing a lag is playback.

The fix is to skip update when the width and height has not changed
from the previously cached size. This reduces calls to MFEngine to only
when there is an actual change in size.

Bug: 1413921
Change-Id: Ied2f6bcf4e658bc6fccf017d8cbc7ac76bd4026d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4233402
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Daoyuan Li <daoyuanli@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1105345}