dns: check for res_ninit before use
The system DNS resolver header on POSIX systems (resolv.h) defines a
macro named __RES, which is set to the API version that the resolver
supports. API version 19991006 is the one which added res_ninit(3) and
res_nclose(3). DnsReloader previously had some bespoke preprocessor
conditionals for this, but what we actually need to check is:

* Are we on a POSIX system? If so...
* Is __RES defined and >= 19991006?
* Are we not on Apple? (https://crbug.com/1414923)
* Are we not on Android?
* Are we not on Fuchsia? (broken system headers)

For comparison, we used to check:

* POSIX && !APPLE && !OPENBSD && !ANDROID && !FUCHSIA (cc file)
* POSIX && !APPLE && !OPENBSD (header)
* POSIX && !APPLE && !OPENBSD && !ANDROID (call site)

This change centralizes the check into one place and makes it more correct, by
defining stub versions of EnsureDnsReloaderInit() and DnsReloaderMaybeReload()
on the platforms that don't support dns reloading.

Bug: 1380654
Change-Id: Ia7d4f5f776db318e6ae60d5bb8d7df6f352e7291
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4242103
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105358}