Prevent password re-authentication dialog from showing over the screensaver on Mac
We previously used the system-defined `system.login.screensaver` access right when authenticating prior to accessing passwords. This right delegates to the `use-login-window-ui` rule. The resulting authentication dialog displays at a window level that appears over the screensaver. Additionally, it allows authentication by either the session owner or any admin user. This is contrary to the expected behavior of permitting only the session owner as described within the code.

We now define a custom access right, `<bundle-id>.access-passwords`, that delegates to kAuthorizationRuleAuthenticateAsSessionUser. This uses a standard window level that does not appear over the screensaver and also limits authentication to the session owner.

Fixed: 1066858
Change-Id: I1e1a486efd65d2cf312d3a726a36f35daeaf39f8
Low-Coverage-Reason: System authentication prompt is challenging to test in an automated fashion.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4241666
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Mark Rowe <markrowe@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105366}