Prefer official paper names for duplicate sizes
In crrev.com/c/4157931, we dropped borderless sizes from the localized
paper name list so they wouldn't override standard size names.  We have
subsequently found printers that report multiple non-borderless paper
entries with the same dimensions.  We can't just drop all of the
non-standard names because sometimes there are no official PWG sizes
corresponding to any of the entries.

What we ideally want is to use the official PWG name that has a curated
localization if it is available for a particular size.  If no official
name for a given size is available, we should pick any of the
non-borderless entries to display.

Achieve this in two steps:

1. Before generating a name for a non-standard size, look it up to see
   if the dimensions correspond to an official size.  Use the same
   display name as the official size if one exists.  This ensures that
   all the entries with the same dimensions sort next to each other.
2. Track whether each entry came from the table of official sizes and
   sort official sizes above non-standard ones.  This means that the
   official size will always be first if it exists (and therefore chosen
   during de-duplication).  If an official size doesn't exist, the first
   non-standard size will be chosen.

Bug: b:268586056
Test: Manual review of paper size lists on lab printers
Change-Id: I3b2604a287f94a0685e80a5860a9ef6296d032c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4242274
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105376}