Fully remove TLS 1.0 and TLS 1.1 from //net
https://chromium-review.googlesource.com/c/chromium/src/+/3968743 has
now survived on stable without incident. Originally, as noted in that CL
description, we believe we've actually accidentally disabled this
versions for even longer.

This CL should have broadly[*] the same behavior, but
implemented differently. Where, before, we would silently clamp the
minimum version to TLS 1.2 at the //net layer, now //net acts as if it
doesn't recognize those constants altogether. I've removed the defines
and the Mojo enum values that would make them accessible at all.
Instead, the one place where it was reachable, extensions, applies them
mapping on its end.

I've retained the SSL_CONNECTION_VERSION_* constants because it is still
possible for those to exist in the disk cache, though there's some code
relating to them that I'll remove in a follow-up CL.

[*] The extensions logic ignores unrecognized values, so the one
change in this CL is we used to always fail if max = tls1.1, and
now we'll just ignore it and act as if max = tls1.3.

Bug: 1376584
Change-Id: Icdb4aa5bd4b9405c900c1fe8e6f055805e3fb582
Fixed: 1271986
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4244521
Reviewed-by: Brendon Tiszka <tiszka@google.com>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105395}