ApkWebApps: Consolidate Lacros and non-Lacros codepaths
ApkWebAppService currently has separate codepaths when webapps are
running in Ash vs Lacros. The Lacros codepath was added more recently,
and has better error handling, as it is built around a model of syncing
the installed apps to the desired state, rather than just reacting to
one-off events as they happen.

This CL removes the old Ash-only codepaths, and uses the state-syncing
codepath for both Ash and Lacros. We now only branch based on Lacros
state at the last moment before apps are installed/uninstalled.

This change simplifies the code and makes it easier to maintain while we
are responsible for both Ash and Lacros web apps.

Bug: 1392571
Change-Id: Ia8508d542ad2bc1f3e4ff3fc09bd2a5d6ed481d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220774
Reviewed-by: Alexander Bolodurin <alexbn@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105430}