[Extensions cleanup] Get action visibility in a cleaner way
GetActionVisibility() is complicated to read. However, we can clean it
by removing the call to IsActionVisibleOnToolbar() which checks the
same values (force visibility and is pinned).

With that, IsActionVisibleOnToolbar() can use GetActionVisibility()
instead of having to re compute the same value.

No functionality changed.

Bug: 1416359
Change-Id: Id63b3d819633a42663aaa410becf725874746c11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4251958
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105431}