flac_fuzzer: Filter out the bad data with different channels
This CL fixes the issue that the number of channels of bad data will not
keep consistent fetched from `MetaCallback`, and the channels fecthed
from `WriteCallback` isn't as same as the channels fetched from
`MetaCallback`. For the first case, we should avoid re-write the
`num_channels_` in `MetaCallback`. For the second case, we should abort
in `WriteCallback`.

BUG=chromium:1416146, chromium:1415158

Test: Manual
Change-Id: I516b699168bb8e0d30d010bf6602d54f7656a98c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4251839
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Hongyu Long <hongyulong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105437}