Move CreateNamedPlatformChannel logic to launcher thread.
The Named Platform Channel code manipulates the command line
but the command_line_ member is logically bound to the
launcher thread.

base::CommandLine will soon dcheck for unsafe accesses such
as these.

This CL moves all accesses of the command line member to the
launcher thread, so that there is no potential for a race.

BUG=1344948

Change-Id: I055cda5bf1c392c803a4e4202219e6dd86a13177
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4250178
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105447}