power: Update Adaptive Charging metrics with metrics for slow charging
This updates several existing Adaptive Charging metrics with metrics for slow charging to allow the evaluation of the effectiveness of the slow charging feature. The AdaptiveChargingMinutesToFull and AdaptiveChargingBatteryPercentageOnUnplug metrics are split into NormalCharging, MixedCharging and SlowCharging to provide insight into the results from each charging case that can occur. Historical data will be kept in the histograms with the original metric names with no charging type suffix. The AdaptiveChargingDelayDelta metric is updated to use 3 hours as the correct delay when the available time for charging is 3 hours or more.

BUG=b:260359049

Change-Id: If799d67044f8baf563aaf57b8a2b94d4e96b4d9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4221153
Commit-Queue: Jolene Loo <joleneloo@google.com>
Reviewed-by: Olivier Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105473}