CR2023: Update focus states for cr-button
This CL adds a new div #background that is charge of background
color. The separate div is needed to 1) apply a combination of 2px
border and background-clip to create a transparent inner border on
focus, and 2) still maintain the same paddings, sizings, and custom
borders on cr-button.

Bug: 1400033
Change-Id: Ib35e52eee68f6e2c4be43a38ee9b59720b5d6c0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4247800
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105517}