inputs: Send correct 'process' key in lacros.
Right now, the 'process' key we send to Lacros is being treated as
'Unidentified', because we haven't filled in the 'key' and 'code'
fields of the key event.

Use the same fake 'process' key event as Ash [1] and add tests to
confirm this behaviour.

[1] https://source.chromium.org/chromium/chromium/src/+/main:ui/base/ime/ash/input_method_ash.cc;l=1072;drc=4ea5c16fe3bd5bb3a947fb56c8ded433e7740c84

BUG=b/265853952

Change-Id: I72dcb0ef61276ae994f9c6bc1c312ce19ca742e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4199490
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Curtis McMullan <curtismcmullan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105523}