Assert CreateStableVideoDecoder() happens on UI thread.
This CL addresses a few unresolved comments from CL:3691922 [1, 2, 3].

One of the unresolved comments is related to the threading model of
RenderProcessHostImpl::CreateStableVideoDecoder():

That method assumes that it's always called on the same sequence since
it binds a mojo::Remote the first time it's called and then uses that
mojo::Remote to make interface calls. It has two call sites:

1) From FramelessMediaInterfaceProxy::CreateVideoDecoder() [4]:

A FramelessMediaInterfaceProxy gets created in [5] and in [6]. The
former instance will never call
render_process_host_->CreateStableVideoDecoder() so it's not relevant
for this discussion. The latter instance is more interesting. Since
FramelessMediaInterfaceProxy::CreateVideoDecoder() is a mojo interface
call, it always gets called on the same sequence. Furthermore,
RenderProcessHostImpl::BindMediaInterfaceProxy() gets registered as a
callback for a UI thread interface [7]. Therefore, it seems reasonable
to assume that RenderProcessHostImpl::BindMediaInterfaceProxy() gets
called on the UI thread which in turn would imply that the
FramelessMediaInterfaceProxy::Add() method also gets called on the UI
thread meaning that FramelessMediaInterfaceProxy::CreateVideoDecoder()
also gets called on the UI thread.

2) From MediaInterfaceProxy::CreateVideoDecoder() [8]:

This method already has a DCHECK(thread_checker_.CalledOnValidThread())
[9]. MediaInterfaceProxy instances are created using the
DocumentUserData framework [10]. Note that
RenderFrameHostImpl::BindMediaInterfaceFactoryReceiver() is added to a
mojo::BinderMap [11]. Another callback that is added to the same map
nearby has a DCHECK_CURRENTLY_ON(BrowserThread::UI) [12]. Therefore, it
seems reasonable to assume that
RenderFrameHostImpl::BindMediaInterfaceFactoryReceiver() gets called on
the UI thread which means what MediaInterfaceProxy::Bind() should always
get called on the UI thread and interface calls (including
MediaInterfaceProxy::CreateVideoDecoder()) are received on the UI
thread.

The conclusion is then that both call sites of
RenderProcessHostImpl::CreateStableVideoDecoder() occur on the UI
thread. That means that the |stable_video_decoder_factory_remote_| that
gets bound in RenderProcessHostImpl::CreateStableVideoDecoder() must be
reset and destroyed on the UI thread. The RenderProcessHostImpl
destructor already has a DCHECK_CURRENTLY_ON(BrowserThread::UI). This CL
adds a DCHECK_CURRENTLY_ON(BrowserThread::UI) to
RenderProcessHostImpl::ResetIPC(). That assumption already seems implied
because ResetIPC() invalidates the weak pointers of
|instance_weak_factory_| and based on e.g., [13], it looks like that
WeakPtr factory is bound to the UI thread and so it must be invalidated
on the UI thread.

Another unresolved comment is related to the raw pointer to the
RenderProcessHostImpl that is stored in FramelessMediaInterfaceProxy. In
case someone in the future adds code to the dtor of
FramelessMediaInterfaceProxy to use the RenderProcessHostImpl, it's
safer to destroy the FramelessMediaInterfaceProxy in ResetIPC() so that
the FramelessMediaInterfaceProxy can't possibly use a partially
destroyed RenderProcessHostImpl.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3691922/comment/0a26c52b_b88145b6/
[2] https://chromium-review.googlesource.com/c/chromium/src/+/3691922/comment/b2c432d8_12103535/
[3] https://chromium-review.googlesource.com/c/chromium/src/+/3691922/comment/e60b61a1_428ebbf6/
[4] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/media/frameless_media_interface_proxy.cc;l=85-86;drc=adac219925ef5d9c9a954d189c2e4b8852a4bbed
[5] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/gpu/gpu_data_manager_impl_private.cc;l=839-840;drc=a1cc370d96a1643b7460ccd553579fa5594c5e14
[6] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/render_process_host_impl.cc;l=2522-2523;drc=4e8596dd9e92c04cf894bebcdf81a631074e0ec8
[7] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/render_process_host_impl.cc;l=2463-2466;drc=4e8596dd9e92c04cf894bebcdf81a631074e0ec8
[8] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/media/media_interface_proxy.cc;l=291-292;drc=adac219925ef5d9c9a954d189c2e4b8852a4bbed
[9] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/media/media_interface_proxy.cc;l=278;drc=adac219925ef5d9c9a954d189c2e4b8852a4bbed
[10] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/render_frame_host_impl.cc;l=10749-10750;drc=e44a1e4415e52424aba251b2729517f952fee9df
[11] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/browser_interface_binders.cc;l=939-941;drc=62911c18831a09c91fda635638c224922c52a6d6
[12] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/browser_interface_binders.cc;l=951;drc=62911c18831a09c91fda635638c224922c52a6d6
[13] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/render_process_host_impl.cc;l=2216-2219;drc=4e8596dd9e92c04cf894bebcdf81a631074e0ec8

Bug: b:195769334
Test: None
Change-Id: Ifca4d2e00ce0c6a663084ccbedba37a449f0a180
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4242380
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105532}