[ios] Remove SyncSetupService error APIs in favor of SyncService
Even before this patch, and starting with https://crrev.com/c/4239517,
syncer::SyncService in components/sync is capable of reporting the
precise error that the UI should display. This logic was previously in
SyncSetupService, which now became a thin layer on top.

In this patch, the API in SyncSetupService is removed entirely and all
callers migrated.

Change-Id: I0998467bbce505de08d18fa9a14306941aabcdd1
Bug: 1412320
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4232277
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105565}