Rework the handling of initial values for paint server attributes
Currently for *GradientAttributes/PatternAttributes, a flag exists for
each field to show if it has been set or not. For fields that are just
pointers to SVG DOM objects, this flag would be redundant if it weren't
for the initial values being set during construction of the attributes
object.

Drop said flags and the initial values, then fill out any unset fields
in the structure at the end of the collection phase, using the values
from the entry element. This avoids unnecessary allocations of objects
that will already be available.

Bug: 1414383
Change-Id: Id8684dca00249071916f1772b1c0c3ba2250c1c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4249577
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1105567}