[Profiles] Add histograms recording the StartupProfileMode
Add histograms to better understand the effect of the
kObserverBasedPostProfileInit experiment on the ProfilePicker.

Current metrics show that the experiment causes the ProfilePicker to be
triggered less often on Windows platform. These histograms may help
understand if this is because a browser window is opened more often, or
if the startup code as a whole is called less.

See go/picker-without-guest-metrics (google-only) for some more details.

Bug: 1150326
Change-Id: I1a147464ce953a71558aa06033e318d6ed63df12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4239864
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105626}