[sync] Fix metric Sync.TrustedVaultJavascriptAddRecoveryMethodSucceeded
The previous implementation always returned failures. This is because,
in Java, Void is not instantiable and the success case always accepted
null. This made the success case indistinguishable from the failures
case, once forwarded to the Consumer.

Instead, the success case needs to be distinguished from the failure
case when dealing with the Promise, and propagate a boolean to the
Consumer.

Bug: 1403195
Change-Id: I81aaf4685b493a6c65997713192f335e5fa7c9dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4250858
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105657}