Introduce FormEventLoggerBase::OnDestroyed() and split metrics function.
Currently FormEventLoggerBase::RecordFunnelAndKeyMetrics() is called
from FormEventLoggerBase's destructor. This makes it impossible to
override the functionality in the derived classes, the address and
credit card form event loggers*.

For kAccount profiles, origin-resolved assistance and correctness key
metrics should be emitted. Ideally, this functionality should live in
the address form logger - not in the base class. The idea is thus to
add virtual hooks to the base class, which the address form event
logger can then use.
This CL makes the necessary clean-ups.

The clean-up is twofold:
- Instead of emitting the metrics from the destructor, a OnDestroyed()
  function is introduced. It should always be called before a form
  event logger is destroyed. This is admittedly not particularly nice,
  but seemingly still the least of all evils.
- RecordFunnelAndKeyMetrics() is split into RecordFunnelMetrics() and
  RecordKeyMetrics(). Follow-up CLs will then split RecordKeyMetrics()
  further and introduce virtual functions like RecordAssistance(),
  which the derived classes can overwrite.

* Since by the destruction order, the derived class is called first,
  and consequently the base class cannot call an override from the
  derived class.

Bug: 1007974
Change-Id: I7836130238e4714a24c63cb1883c246bcf7f80c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4249578
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1105662}